split logic into darwin, linux, windows files since only implementing for drawing now, uint32 parsing issue, feedback
