+++ pkg/systray/runtime/runner_linux.go
@@ -1,0 +1,16 @@
+//go:build linux
+// +build linux
+
+package runtime
+
+import (
+	"fmt"
+)
+
+func (r *SystrayUsersProcessesRunner) runConsoleUserSystray() error {
+	return fmt.Errorf("not implemented")
+}
+
+func processExists(pid int) bool {
+	return false
+}
